// telegram-types.ts
export interface Update {
  /**
   * The update's unique identifier. Update identifiers start from a certain 
   * positive number and increase sequentially. This ID becomes especially 
   * handy if you're using Webhooks, since it allows you to ignore repeated 
   * updates or to restore the correct update sequence, should they get out 
   * of order. If there are no new updates for at least a week, then 
   * identifier of the next update will be chosen randomly instead of sequentially.
   * @see https://core.telegram.org/bots/api#setwebhook
   */
  update_id: number;

  /**
   * New incoming message of any kind — text, photo, sticker, etc.
   */
  message?: Message;

  /**
   * New version of a message that is known to the bot and was edited
   */
  edited_message?: Message;

  /**
   * New incoming channel post of any kind — text, photo, sticker, etc.
   */
  channel_post?: Message;

  /**
   * New version of a channel post that is known to the bot and was edited
   */
  edited_channel_post?: Message;

  /**
   * New incoming inline query
   * @see https://core.telegram.org/bots/api#inline-mode
   */
  inline_query?: InlineQuery;

  /**
   * The result of an inline query that was chosen by a user and sent to 
   * their chat partner. Please see our documentation on the feedback 
   * collecting for details on how to enable these updates for your bot.
   * @see https://core.telegram.org/bots/api#inline-mode
   * @see https://core.telegram.org/bots/api/bots/inline#collecting-feedback
   */
  chosen_inline_result?: ChosenInlineResult;

  /**
   * New incoming callback query
   */
  callback_query?: CallbackQuery;

  /**
   * New incoming shipping query. Only for invoices with flexible price
   */
  shipping_query?: ShippingQuery;

  /**
   * New incoming pre-checkout query. Contains full information about checkout
   */
  pre_checkout_query?: PreCheckoutQuery;

  /**
   * New poll state. Bots receive only updates about stopped polls and polls, 
   * which are sent by the bot
   */
  poll?: Poll;

  /**
   * A user changed their answer in a non-anonymous poll. Bots receive new 
   * votes only in polls that were sent by the bot itself.
   */
  poll_answer?: PollAnswer;
}
// ...
export interface MessageEntity {
  /**
   * Type of the entity. Can be “mention” (@username), “hashtag” (#hashtag), 
   * “cashtag” ($USD), “bot_command” (/start@jobs_bot), “url” 
   * (https://telegram.org), “email” (do-not-reply@telegram.org), 
   * “phone_number” (+1-212-555-0123), “bold” (bold text), “italic” (italic 
   * text), “underline” (underlined text), “strikethrough” (strikethrough 
   * text), “code” (monowidth string), “pre” (monowidth block), “text_link” 
   * (for clickable text URLs), “text_mention” (for users without usernames)
   * @see https://telegram.org/blog/edit#new-mentions
   */
  type: string;

  /**
   * Offset in UTF-16 code units to the start of the entity
   */
  offset: number;

  /**
   * Length of the entity in UTF-16 code units
   */
  length: number;

  /**
   * For “text_link” only, url that will be opened after user taps on the text
   */
  url?: string;

  /**
   * For “text_mention” only, the mentioned user
   */
  user?: User;

  /**
   * For “pre” only, the programming language of the entity text
   */
  language?: string;
}
// ...
export interface PollOption {
  /**
   * Option text, 1-100 characters
   */
  text: string;

  /**
   * Number of users that voted for this option
   */
  voter_count: number;
}

/**
 * This object represents an answer of a user in a non-anonymous poll.
 */
export interface PollAnswer {
  /**
   * Unique poll identifier
   */
  poll_id: string;

  /**
   * The user, who changed the answer to the poll
   */
  user: User;

  /**
   * 0-based identifiers of answer options, chosen by the user. May be empty 
   * if the user retracted their vote.
   */
  option_ids: number[];
}

/**
 * This object contains information about a poll.
 */
export interface Poll {
  /**
   * Unique poll identifier
   */
  id: string;

  /**
   * Poll question, 1-300 characters
   */
  question: string;

  /**
   * List of poll options
   */
  options: PollOption[];

  /**
   * Total number of users that voted in the poll
   */
  total_voter_count: number;

  /**
   * True, if the poll is closed
   */
  is_closed: boolean;

  /**
   * True, if the poll is anonymous
   */
  is_anonymous: boolean;

  /**
   * Poll type, currently can be “regular” or “quiz”
   */
  type: string;

  /**
   * True, if the poll allows multiple answers
   */
  allows_multiple_answers: boolean;

  /**
   * 0-based identifier of the correct answer option. Available only for 
   * polls in the quiz mode, which are closed, or was sent (not forwarded) by 
   * the bot or to the private chat with the bot.
   */
  correct_option_id?: number;

  /**
   * Text that is shown when a user chooses an incorrect answer or taps on 
   * the lamp icon in a quiz-style poll, 0-200 characters
   */
  explanation?: string;

  /**
   * Special entities like usernames, URLs, bot commands, etc. that appear in 
   * the explanation
   */
  explanation_entities?: MessageEntity[];

  /**
   * Amount of time in seconds the poll will be active after creation
   */
  open_period?: number;

  /**
   * Point in time (Unix timestamp) when the poll will be automatically closed
   */
  close_date?: number;
}
// ...
export interface InlineKeyboardMarkup {
  /**
   * Array of button rows, each represented by an Array of 
   * InlineKeyboardButton objects
   * @see https://core.telegram.org/bots/api#inlinekeyboardbutton
   */
  inline_keyboard: InlineKeyboardButton[][];
}
// ...
export interface CallbackQuery {
  /**
   * Unique identifier for this query
   */
  id: string;

  /**
   * Sender
   */
  from: User;

  /**
   * Message with the callback button that originated the query. Note that 
   * message content and message date will not be available if the message is 
   * too old
   */
  message?: Message;

  /**
   * Identifier of the message sent via the bot in inline mode, that 
   * originated the query.
   */
  inline_message_id?: string;

  /**
   * Global identifier, uniquely corresponding to the chat to which the 
   * message with the callback button was sent. Useful for high scores in games.
   * @see https://core.telegram.org/bots/api#games
   */
  chat_instance: string;

  /**
   * Data associated with the callback button. Be aware that a bad client can 
   * send arbitrary data in this field.
   */
  data?: string;

  /**
   * Short name of a Game to be returned, serves as the unique identifier for 
   * the game
   * @see https://core.telegram.org/bots/api#games
   */
  game_short_name?: string;
}
// ...
export interface SendMessage {
  /**
   * Unique identifier for the target chat or username of the target channel 
   * (in the format @channelusername)
   */
  chat_id: (number | string);

  /**
   * Text of the message to be sent, 1-4096 characters after entities parsing
   */
  text: string;

  /**
   * Mode for parsing entities in the message text. See formatting options 
   * for more details.
   * @see https://core.telegram.org/bots/api#formatting-options
   */
  parse_mode?: string;

  /**
   * List of special entities that appear in message text, which can be 
   * specified instead of parse_mode
   */
  entities?: MessageEntity[];

  /**
   * Disables link previews for links in this message
   */
  disable_web_page_preview?: boolean;

  /**
   * Sends the message silently. Users will receive a notification with no sound.
   * @see https://telegram.org/blog/channels-2-0#silent-messages
   */
  disable_notification?: boolean;

  /**
   * If the message is a reply, ID of the original message
   */
  reply_to_message_id?: number;

  /**
   * Pass True, if the message should be sent even if the specified 
   * replied-to message is not found
   */
  allow_sending_without_reply?: boolean;

  /**
   * Additional interface options. A JSON-serialized object for an inline 
   * keyboard, custom reply keyboard, instructions to remove reply keyboard 
   * or to force a reply from the user.
   * @see https://core.telegram.org/bots#inline-keyboards-and-on-the-fly-updating
   * @see https://core.telegram.org/bots#keyboards
   */
  reply_markup?: (InlineKeyboardMarkup | ReplyKeyboardMarkup | ReplyKeyboardRemove | ForceReply);
}
// ...
export interface EditMessageText {
  /**
   * Required if inline_message_id is not specified. Unique identifier for 
   * the target chat or username of the target channel (in the format @channelusername)
   */
  chat_id?: (number | string);

  /**
   * Required if inline_message_id is not specified. Identifier of the 
   * message to edit
   */
  message_id?: number;

  /**
   * Required if chat_id and message_id are not specified. Identifier of the 
   * inline message
   */
  inline_message_id?: string;

  /**
   * New text of the message, 1-4096 characters after entities parsing
   */
  text: string;

  /**
   * Mode for parsing entities in the message text. See formatting options 
   * for more details.
   * @see https://core.telegram.org/bots/api#formatting-options
   */
  parse_mode?: string;

  /**
   * List of special entities that appear in message text, which can be 
   * specified instead of parse_mode
   */
  entities?: MessageEntity[];

  /**
   * Disables link previews for links in this message
   */
  disable_web_page_preview?: boolean;

  /**
   * A JSON-serialized object for an inline keyboard.
   * @see https://core.telegram.org/bots#inline-keyboards-and-on-the-fly-updating
   */
  reply_markup?: InlineKeyboardMarkup;
}
// ...
export interface InlineQuery {
  /**
   * Unique identifier for this query
   */
  id: string;

  /**
   * Sender
   */
  from: User;

  /**
   * Sender location, only for bots that request user location
   */
  location?: Location;

  /**
   * Text of the query (up to 256 characters)
   */
  query: string;

  /**
   * Offset of the results to be returned, can be controlled by the bot
   */
  offset: string;
}
// ...
export interface ChosenInlineResult {
  /**
   * The unique identifier for the result that was chosen
   */
  result_id: string;

  /**
   * The user that chose the result
   */
  from: User;

  /**
   * Sender location, only for bots that require user location
   */
  location?: Location;

  /**
   * Identifier of the sent inline message. Available only if there is an 
   * inline keyboard attached to the message. Will be also received in 
   * callback queries and can be used to edit the message.
   * @see https://core.telegram.org/bots/api#inlinekeyboardmarkup
   * @see https://core.telegram.org/bots/api#callbackquery
   * @see https://core.telegram.org/bots/api#updating-messages
   */
  inline_message_id?: string;

  /**
   * The query that was used to obtain the result
   */
  query: string;
}
// ...
export interface ShippingQuery {
  /**
   * Unique query identifier
   */
  id: string;

  /**
   * User who sent the query
   */
  from: User;

  /**
   * Bot specified invoice payload
   */
  invoice_payload: string;

  /**
   * User specified shipping address
   */
  shipping_address: ShippingAddress;
}

/**
 * This object contains information about an incoming pre-checkout query.
 */
export interface PreCheckoutQuery {
  /**
   * Unique query identifier
   */
  id: string;

  /**
   * User who sent the query
   */
  from: User;

  /**
   * Three-letter ISO 4217 currency code
   * @see https://core.telegram.org/bots/api/bots/payments#supported-currencies
   */
  currency: string;

  /**
   * Total price in the smallest units of the currency (integer, not 
   * float/double). For example, for a price of US$ 1.45 pass amount = 145. 
   * See the exp parameter in currencies.json, it shows the number of digits 
   * past the decimal point for each currency (2 for the majority of currencies).
   * @see https://core.telegram.org/bots/payments/currencies.json
   */
  total_amount: number;

  /**
   * Bot specified invoice payload
   */
  invoice_payload: string;

  /**
   * Identifier of the shipping option chosen by the user
   */
  shipping_option_id?: string;

  /**
   * Order info provided by the user
   */
  order_info?: OrderInfo;
}